backend/apps/integrations/confluence/client.py
import logging

from atlassian import Confluence

logger = logging.getLogger(__name__)


class ConfluenceClient:
    """Client für die Kommunikation mit der Confluence REST API."""

    def __init__(self, url: str, email: str, api_token: str):
        self.url = url
        self.email = email
        self.confluence = Confluence(
            url=url,
            username=email,
            password=api_token,
            cloud=True,
        )

    def get_spaces(self) -> list[dict]:
        """Alle verfügbaren Spaces abrufen."""
        # TODO: Paginierung implementieren
        # TODO: Berechtigungen beachten
        raise NotImplementedError

    def get_pages(self, space_key: str) -> list[dict]:
        """Seiten eines Spaces abrufen."""
        # TODO: Paginierung implementieren
        # TODO: Nur geänderte Seiten seit letztem Sync laden
        raise NotImplementedError

    def get_page_content(self, page_id: str) -> dict:
        """Inhalt einer einzelnen Seite abrufen."""
        # TODO: HTML-Content in Plain Text konvertieren
        # TODO: Anhänge und eingebettete Medien berücksichtigen
        raise NotImplementedError

    def create_page(self, space_key: str, title: str, body: str, parent_id: str = None) -> dict:
        """Neue Seite in Confluence erstellen."""
        # TODO: Validierung der Eingabedaten
        # TODO: HTML-Formatierung des Body
        raise NotImplementedError

    def update_page(self, page_id: str, title: str, body: str) -> dict:
        """Bestehende Seite aktualisieren."""
        # TODO: Versionsnummer für optimistic locking abrufen
        # TODO: Konflikterkennung implementieren
        raise NotImplementedError

    def get_page_history(self, page_id: str) -> list[dict]:
        """Änderungshistorie einer Seite abrufen."""
        # TODO: Implementierung mit Paginierung
        raise NotImplementedError
