backend/requirements/base.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
redis>=5.0,<6.0
atlassian-python-api>=3.41,<4.0
msal>=1.28,<2.0
httpx>=0.27,<1.0
anthropic>=0.34,<1.0
python-decouple>=3.8,<4.0
drf-spectacular>=0.27,<1.0
gunicorn>=22.0,<23.0
